+++ drools-core/src/main/java/org/drools/core/phreak/PhreakExistsNode.java
@@ -27,6 +27,7 @@ import org.drools.core.reteoo.TupleMemory;
 import org.drools.core.rule.ContextEntry;
 import org.drools.core.spi.PropagationContext;
 import org.drools.core.util.FastIterator;
+import org.drools.core.util.PerfLogUtils;
 
 import static org.drools.core.phreak.PhreakJoinNode.updateChildLeftTuple;
 
@@ -45,6 +46,8 @@ public class PhreakExistsNode {
                        TupleSets<LeftTuple> srcLeftTuples,
                        TupleSets<LeftTuple> trgLeftTuples,
                        TupleSets<LeftTuple> stagedLeftTuples) {
+
+        PerfLogUtils.startMetrics(existsNode);
 
         TupleSets<RightTuple> srcRightTuples = bm.getStagedRightTuples().takeAll();
 
@@ -90,6 +93,8 @@ public class PhreakExistsNode {
 
         srcRightTuples.resetAll();
         srcLeftTuples.resetAll();
+
+        PerfLogUtils.logAndEndMetrics();
     }
 
     public void doLeftInserts(ExistsNode existsNode,
@@ -161,6 +166,7 @@ public class PhreakExistsNode {
                     }
 
                     // we know that only unblocked LeftTuples are  still in the memory
+                    PerfLogUtils.incrementEvalCount();
                     if ( constraints.isAllowedCachedRight( contextEntry,
                                                            leftTuple ) ) {
                         leftTuple.setBlocker( rightTuple );
@@ -228,6 +234,7 @@ public class PhreakExistsNode {
             }
 
             // if we where not blocked before (or changed buckets), or the previous blocker no longer blocks, then find the next blocker
+            PerfLogUtils.incrementEvalCount();
             if (blocker == null || !constraints.isAllowedCachedLeft(contextEntry,
                                                                     blocker.getFactHandleForEvaluation())) {
 
@@ -238,6 +245,7 @@ public class PhreakExistsNode {
 
                 // find first blocker, because it's a modify, we need to start from the beginning again
                 for (RightTuple newBlocker = firstRightTuple; newBlocker != null; newBlocker = (RightTuple) rightIt.next(newBlocker)) {
+                    PerfLogUtils.incrementEvalCount();
                     if (constraints.isAllowedCachedLeft( contextEntry,
                                                          newBlocker.getFactHandleForEvaluation() )) {
                         leftTuple.setBlocker( newBlocker );
@@ -318,6 +326,7 @@ public class PhreakExistsNode {
                     }
 
                     // we know that only unblocked LeftTuples are  still in the memory
+                    PerfLogUtils.incrementEvalCount();
                     if ( constraints.isAllowedCachedRight( contextEntry,
                                                            leftTuple ) ) {
                         leftTuple.setBlocker( rightTuple );
@@ -371,6 +380,7 @@ public class PhreakExistsNode {
                     for ( RightTuple newBlocker = rootBlocker; newBlocker != null; newBlocker = (RightTuple) rightIt.next( newBlocker ) ) {
                         // cannot select a RightTuple queued in the delete list
                         // There may be UPDATE RightTuples too, but that's ok. They've already been re-added to the correct bucket, safe to be reprocessed.
+                        PerfLogUtils.incrementEvalCount();
                         if ( leftTuple.getStagedType() != LeftTuple.DELETE && newBlocker.getStagedType() != LeftTuple.DELETE &&
                              constraints.isAllowedCachedLeft( contextEntry, newBlocker.getFactHandleForEvaluation() ) ) {
                             leftTuple.setBlocker( newBlocker );
@@ -476,6 +486,7 @@ public class PhreakExistsNode {
 
                     // we know that older tuples have been checked so continue previously
                     for (RightTuple newBlocker = rootBlocker; newBlocker != null; newBlocker = (RightTuple) it.next(newBlocker)) {
+                        PerfLogUtils.incrementEvalCount();
                         if (!newBlocker.isDeleted() && constraints.isAllowedCachedLeft(contextEntry,
                                                                                        newBlocker.getFactHandleForEvaluation())) {
                             leftTuple.setBlocker(newBlocker);
